♻️ Move rows.go to internal/wrap
